Add images to challah blog post
I'm not sure if it will make sense to keep storing images in this repo,
but it's easy for now.

I put the images in the directory where I wanted them and then ran

```
sips -Z 960 assets/**/*.jpg
```

which is a mac utility to resize and make basic manipulations to image
files. The -Z tells it to maintain aspect ratio.

Eventually I want to set up one of the many jekyll plugins for
responsive images because that sounds like the right thing to do and I
want to know how to do it.